+++ packages/app/src/hooks/useCreatePool/util.ts
@@ -1,0 +1,4 @@
+// TODO Upload image
+const uploadImg = async (img: File, name: string) => {};
+
+export { uploadImg };
